separate out sort config
