+++ web-common/src/features/canvas/inspector/chart/field-config/FieldConfigPopover.svelte
@@ -1,10 +1,8 @@
 <script lang="ts">
   import IconButton from "@rilldata/web-common/components/button/IconButton.svelte";
-  import DraggableList from "@rilldata/web-common/components/draggable-list";
   import Input from "@rilldata/web-common/components/forms/Input.svelte";
   import Select from "@rilldata/web-common/components/forms/Select.svelte";
   import Switch from "@rilldata/web-common/components/forms/Switch.svelte";
-  import DragHandle from "@rilldata/web-common/components/icons/DragHandle.svelte";
   import SettingsSlider from "@rilldata/web-common/components/icons/SettingsSlider.svelte";
   import * as Popover from "@rilldata/web-common/components/popover";
   import type {
@@ -12,7 +10,7 @@
     FieldConfig,
   } from "@rilldata/web-common/features/canvas/components/charts/types";
   import type { ChartFieldInput } from "@rilldata/web-common/features/canvas/inspector/types";
-  import { List } from "lucide-svelte";
+  import SortConfig from "./SortConfig.svelte";
 
   export let fieldConfig: FieldConfig;
   export let onChange: (property: keyof FieldConfig, value: any) => void;
@@ -29,15 +27,6 @@
   let labelAngle =
     fieldConfig?.labelAngle ?? (fieldConfig?.type === "temporal" ? 0 : -90);
   let isDropdownOpen = false;
-  let isCustomSortDropdownOpen = false;
-
-  const sortOptions = [
-    { label: "Ascending", value: "x" },
-    { label: "Descending", value: "-x" },
-    { label: "Y-axis ascending", value: "y" },
-    { label: "Y-axis descending", value: "-y" },
-    { label: "Custom", value: "custom" },
-  ];
 
   const legendOptions: { label: string; value: ChartLegend }[] = [
     { label: "Top", value: "top" },
@@ -55,37 +44,6 @@
   $: showLabelAngle = chartFieldInput?.labelAngleSelector ?? false;
   $: showLegend = chartFieldInput?.defaultLegendOrientation ?? false;
   $: showAxisRange = chartFieldInput?.axisRangeSelector ?? false;
-
-  $: sortValue = fieldConfig.sort
-    ? typeof fieldConfig.sort === "string"
-      ? fieldConfig.sort
-      : "custom"
-    : (sortConfig.defaultSort ?? "x");
-
-  $: customSortDraggableItems = sortConfig.customSortItems?.map((item) => ({
-    id: item,
-    value: item,
-  }));
-
-  function handleReorder(
-    event: CustomEvent<{
-      items: { id: string; value: string }[];
-      fromIndex: number;
-      toIndex: number;
-    }>,
-  ) {
-    const reorderedItems = event.detail.items.map((item) => item.value);
-    onChange("sort", reorderedItems);
-  }
-
-  function handleSortChange(e: CustomEvent<string>) {
-    const newSortValue = e.detail;
-    if (newSortValue === "custom") {
-      onChange("sort", sortConfig.customSortItems);
-    } else {
-      onChange("sort", newSortValue);
-    }
-  }
 </script>
 
 <Popover.Root bind:open={isDropdownOpen}>
@@ -138,54 +96,7 @@
             />
           </div>
         {/if}
-        {#if sortConfig.enable}
-          <div class="py-1 flex items-center justify-between">
-            <span class="text-xs">Sort</span>
-            <div class="flex items-center gap-x-1">
-              <Select
-                size="sm"
-                id="sort-select"
-                width={190}
-                options={sortOptions}
-                value={sortValue}
-                on:change={handleSortChange}
-              />
-              {#if sortValue === "custom"}
-                <Popover.Root bind:open={isCustomSortDropdownOpen}>
-                  <Popover.Trigger>
-                    <IconButton rounded active={isCustomSortDropdownOpen}>
-                      <List size="14px" />
-                    </IconButton>
-                  </Popover.Trigger>
-                  <Popover.Content align="end" class="w-[240px] p-0">
-                    <div class="px-3 py-2 border-b border-gray-200">
-                      <span class="text-xs font-medium">Sort Order</span>
-                    </div>
-                    <DraggableList
-                      items={customSortDraggableItems || []}
-                      on:reorder={handleReorder}
-                      minHeight="auto"
-                      maxHeight="300px"
-                    >
-                      <div slot="empty" class="px-2 py-2 text-xs text-gray-500">
-                        No sort item found
-                      </div>
-                      <div
-                        slot="item"
-                        let:item
-                        class="flex items-center gap-x-1"
-                      >
-                        <DragHandle size="16px" className="text-gray-400" />
-
-                        <span class="text-xs truncate">{item.value}</span>
-                      </div>
-                    </DraggableList>
-                  </Popover.Content>
-                </Popover.Root>
-              {/if}
-            </div>
-          </div>
-        {/if}
+        <SortConfig {fieldConfig} {onChange} {sortConfig} />
         {#if showLimit}
           <div class="py-1 flex items-center justify-between">
             <span class="text-xs">Limit</span>
